calibration: move force calibration items
This renames `ForceCalibration` to `ForceCalibrationList` and moves the `ForceCalibrationItem` (which is the `Marker`-like item) to a more appropriate location.
