+++ lumicks/pylake/force_calibration/tests/ref_data/calibration_ac.json
@@ -3,7 +3,7 @@
   "parametrization": {},
   "result": {
     "num_points_per_block": 200,
-    "sample_rate": 78125.0,
+    "sample_rate": 78125,
     "excluded_ranges": [
       [
         12.0,
