+++ lumicks/pylake/force_calibration/tests/ref_data/model.json
@@ -2,22 +2,6 @@
   "test": "passive_item",
   "parametrization": {},
   "result": {
-    "num_points_per_block": 2000,
-    "sample_rate": 78125.0,
-    "excluded_ranges": [
-      [
-        100.0,
-        500.0
-      ],
-      [
-        1000.0,
-        2000.0
-      ]
-    ],
-    "fit_range": [
-      100.0,
-      23000.0
-    ],
     "bead_diameter": 1.0,
     "rho_bead": 1100.0,
     "rho_sample": 1000.0,
@@ -26,9 +10,9 @@
     "distance_to_surface": 1.0,
     "fixed_alpha": 0.45,
     "fixed_diode": 10000.0,
+    "sample_rate": 78125,
     "fast_sensor": false,
     "axial": false,
-    "hydrodynamically_correct": true,
-    "active_calibration": false
+    "hydrodynamically_correct": true
   }
 }
