+++ lumicks/pylake/tests/test_file/conftest.py
@@ -254,12 +254,3 @@ def h5_two_colors(tmpdir_factory, request):
     mock_file.make_continuous_channel("Photon count", "Blue", np.int64(20e9), freq, counts)
     mock_file.make_continuous_channel("Info wave", "Info wave", np.int64(20e9), freq, infowave)
     return mock_file.file
-
-
-@pytest.fixture()
-def calibration_data():
-    np.random.seed(1337)
-    num_samples = 100000
-    dummy_voltage = np.random.normal(size=num_samples)
-    dummy_nano = np.sin(2.0 * np.pi * 17 * np.arange(num_samples) / 78125)
-    return dummy_voltage, dummy_nano
