+++ lumicks/pylake/tests/test_file/test_file_items.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from lumicks import pylake
-from lumicks.pylake.calibration import ForceCalibration
+from lumicks.pylake.calibration import ForceCalibrationList
 
 from ..data.mock_json import force_feedback_dict
 
@@ -33,9 +33,9 @@ def test_channels(h5_file):
 def test_calibration(h5_file):
     f = pylake.File.from_h5py(h5_file)
 
-    assert type(f.force1x.calibration) is ForceCalibration or list
-    assert type(f.downsampled_force1.calibration) is ForceCalibration or list
-    assert type(f.downsampled_force1x.calibration) is ForceCalibration or list
+    assert type(f.force1x.calibration) is ForceCalibrationList or list
+    assert type(f.downsampled_force1.calibration) is ForceCalibrationList or list
+    assert type(f.downsampled_force1x.calibration) is ForceCalibrationList or list
 
     if f.format_version == 1:
         # v1 version doesn't include calibration data field
